test_fastapi/pyproject.toml
[project]
name = "test-fastapi"
version = "0.1.0"
description = "Test application for fastapi-role"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "casbin",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "uvicorn",
    "platformdirs",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist",
    "pytest-sugar",
    "pytest-clarity",
    "pytest-icdiff",
    "pytest-emoji",
    "ruff",
    "mypy",
    "httpx",
    "hypothesis",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["test_fastapi"]

[tool.uv.sources]
fastapi-role = { workspace = true }
